+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 - Classes to create Material icons from `span`s to add icons to Markdown.
 - Property to `GlobalStyle` for adding custom styles.
+- `business` icon to `Icon`.
 
 ## [0.12.0] - 2022-01-29