+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - `business` icon to `Icon`.
 - `cloud` icon to `Icon`.
 - Property to `Icon` to modify the style of the icons.
+- `download` icon to `Icon`.
+- `file_download` icon to `Icon`.
 
 ### Changed