+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## Unreleased
+## [0.13.0] - 2022-02-27
 
 ### Added
 
@@ -140,7 +140,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Card component.
 - Material icon component.
 
-[unreleased]: https://github.com/visiosto/react-components/compare/v0.12.0...HEAD
+[unreleased]: https://github.com/visiosto/react-components/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/visiosto/react-components/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/visiosto/react-components/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/visiosto/react-components/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/visiosto/react-components/compare/v0.9.1...v0.10.0